# api/openapi-v1.yml
openapi: 3.0.3
info:
  title: Thesis Discovery Platform API
  version: 0.1.0
  license:
    name: Unlicensed
    url: https://choosealicense.com/no-permission/
  description: |
    Simplified API contract for the thesis discovery platform.

    The system has four services:

    - Thesis Service: owns PostgreSQL data and exposes the frontend API.
    - Scraping Service: periodically scrapes chair/source websites and sends extracted results to the Thesis Service.
    - GenAI Service: converts raw scraped content into structured thesis entries.
    - Vector Search Service: owns the vector database and supports semantic search plus reindexing.

    Important design decision:
    During every successful scrape run for a chair, the Thesis Service replaces all existing thesis proposals
    for that chair in one transaction before instructing the Vector Search Service to refresh the
    corresponding vector index entries.
security: []
servers:
  - url: http://localhost:8080
    description: Thesis Service
  - url: http://localhost:8081
    description: Scraping Service
  - url: http://localhost:8082
    description: Vector Search Service
  - url: http://localhost:8000
    description: GenAI Service
...
paths:
...
  /internal/v1/genai-service/extract-theses:
    post:
      tags:
        - GenAI API
      summary: Extract structured theses from scraped content
      description: |
        Called by the Scraping Service after it downloads raw HTML/text from a source endpoint.

        The GenAI Service returns structured thesis proposals extracted from the raw content.
        It does not write to PostgreSQL or the vector database.
      operationId: extractThesesFromRawContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenAIExtractionRequest'
            example:
              sourceEndpointId: 10
              chairId: 3
              chairName: "Chair of Software Engineering"
              sourceUrl: "https://example-chair.tum.de/theses"
              rawHtml: "<html>...</html>"
              knownResearchAreas: ["Artificial Intelligence", "Robotics", "Security"]
      responses:
        '200':
          description: Structured thesis proposals extracted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenAIExtractionResponse'
        '422':
          description: Extraction could not produce reliable structured output.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
...
  schemas:
...
    ThesisProposalInput:
      type: object
      required:
        - title
        - sourceUrl
      properties:
        title:
          type: string
        degreeType:
          type: string
          nullable: true
        originalDescription:
          type: string
          nullable: true
        aiOverview:
          type: string
          nullable: true
        researchArea:
          type: string
          nullable: true
        sourceUrl:
          type: string
          format: uri
        status:
          type: string
          default: "OPEN"
        advisors:
          type: array
          items:
            $ref: '#/components/schemas/AdvisorInput'
...
    AdvisorInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        profileUrl:
          type: string
          format: uri
          nullable: true
...
    GenAIExtractionRequest:
      type: object
      required:
        - sourceEndpointId
        - chairId
        - sourceUrl
        - rawHtml
      properties:
        sourceEndpointId:
          type: integer
          format: int64
        chairId:
          type: integer
          format: int64
        chairName:
          type: string
          nullable: true
        sourceUrl:
          type: string
          format: uri
        rawHtml:
          type: string
        extractedPlainText:
          type: string
          nullable: true
          description: Optional plain text extracted by the Scraping Service before calling GenAI.
        knownResearchAreas:
          type: array
          description: Existing/canonical research area names the model should reuse whenever possible.
          items:
            type: string
    GenAIExtractionResponse:
      type: object
      required:
        - theses
      properties:
        theses:
          type: array
          items:
            $ref: '#/components/schemas/ThesisProposalInput'
        extractionNotes:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
...